requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0

# Celery for Background Tasks
celery>=5.3.0
redis>=4.5.0

# Database
psycopg2-binary>=2.9.0

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.0
librosa>=0.10.0

# Environment Management
python-decouple>=3.8
django-environ>=0.10.0
dj-database-url>=2.0.0

# File Storage
django-storages>=1.13.0
boto3>=1.26.0

# Development
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
pytest>=7.0.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.0.0
whitenoise>=6.0.0
sentry-sdk>=1.25.0